move createEmailTemplate to mail.ts && delete user acc if email fails to reach mail service
